moves order-by-order bernoulli functions out of the pq_helper class
