+++ pdaggerq/pq_helper.h
@@ -293,78 +293,6 @@ class pq_helper {
                                                                 const std::vector<std::string> &targets,
                                                                 const std::vector<std::string> &ops,
                                                                 const int max_order);
-    /**
-     *
-     * generate list of first-order terms from the Bernoulli-number representation of the similarity-transformed operator
-     *
-     * @param targets: a list of strings defining the operator product to be transformed (here, f)
-     * @param ops: a list of strings defining a sum of operators that define the transformation (here, T)
-     *
-     */
-    std::vector<pq_operator_terms> get_bernoulli_operator_terms_1(double factor, 
-                                                                  const std::vector<std::string> &targets,
-                                                                  const std::vector<std::string> &ops);
-
-    /**
-     *
-     * generate list of second-order terms from the Bernoulli-number representation of the similarity-transformed operator
-     *
-     * @param targets: a list of strings defining the operator product to be transformed (here, f)
-     * @param ops: a list of strings defining a sum of operators that define the transformation (here, T)
-     *
-     */
-    std::vector<pq_operator_terms> get_bernoulli_operator_terms_2(double factor, 
-                                                                  const std::vector<std::string> &targets,
-                                                                  const std::vector<std::string> &ops);
-
-    /**
-     *
-     * generate list of third-order terms from the Bernoulli-number representation of the similarity-transformed operator
-     *
-     * @param targets: a list of strings defining the operator product to be transformed (here, f)
-     * @param ops: a list of strings defining a sum of operators that define the transformation (here, T)
-     *
-     */
-    std::vector<pq_operator_terms> get_bernoulli_operator_terms_3(double factor, 
-                                                                  const std::vector<std::string> &targets,
-                                                                  const std::vector<std::string> &ops);
-
-    /**
-     *
-     * generate list of fourth-order terms from the Bernoulli-number representation of the similarity-transformed operator
-     *
-     * @param targets: a list of strings defining the operator product to be transformed (here, f)
-     * @param ops: a list of strings defining a sum of operators that define the transformation (here, T)
-     *
-     */
-    std::vector<pq_operator_terms> get_bernoulli_operator_terms_4(double factor, 
-                                                                  const std::vector<std::string> &targets,
-                                                                  const std::vector<std::string> &ops);
-
-    /**
-     *
-     * generate list of fifth-order terms from the Bernoulli-number representation of the similarity-transformed operator
-     *
-     * @param targets: a list of strings defining the operator product to be transformed (here, f)
-     * @param ops: a list of strings defining a sum of operators that define the transformation (here, T)
-     *
-     */
-    std::vector<pq_operator_terms> get_bernoulli_operator_terms_5(double factor, 
-                                                                  const std::vector<std::string> &targets,
-                                                                  const std::vector<std::string> &ops);
-
-    /**
-     *
-     * generate list of sixth-order terms from the Bernoulli-number representation of the similarity-transformed operator
-     *
-     * @param targets: a list of strings defining the operator product to be transformed (here, f)
-     * @param ops: a list of strings defining a sum of operators that define the transformation (here, T)
-     *
-     */
-    std::vector<pq_operator_terms> get_bernoulli_operator_terms_6(double factor, 
-                                                                  const std::vector<std::string> &targets,
-                                                                  const std::vector<std::string> &ops);
-
 
     /**
      *
